requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
cloudevents==1.9.0
deprecation==2.1.0
Flask==2.2.2
frozenlist==1.3.3
functions-framework==3.3.0
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.4
numpy==1.24.1
openai==0.27.0
packaging==22.0
pandas==1.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7
requests==2.28.1
six==1.16.0
tqdm==4.65.0
urllib3==1.26.13
watchdog==2.2.1
Werkzeug==2.2.2
yarl==1.8.2
